update: bigint

+++ book/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Book(models.Model):
-    number = models.BigIntegerField(null=True, blank=True)
+    number = models.IntegerField(null=True, blank=True)
     title = models.TextField(null=True, blank=True)
     author = models.TextField(null=True, blank=True)
     rating = models.FloatField(null=True, blank=True)
@@ -13,7 +13,7 @@ class Book(models.Model):
     publisher = models.TextField(null=True, blank=True)
     page_count = models.IntegerField(null=True, blank=True)
     generes = models.TextField(null=True, blank=True)
-    ISBN = models.IntegerField(null=True, blank=True)
+    ISBN = models.BigIntegerField(null=True, blank=True)
     language = models.TextField(null=True, blank=True)
     published_date = models.TextField(null=True, blank=True)
     image = models.TextField(null=True, blank=True)
